fix: move batchQuery reducer auxiliar file to utils folder
