+++ utils/batchQuery/reducer.ts
